# docs/swagger.yaml
basePath: /
definitions:
...
  entities.Role:
    properties:
      created_at:
        type: string
      domain_id:
        type: string
      id:
        type: string
      role_claims:
        additionalProperties: true
        type: object
      role_name:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.CreateRoleRequest:
    properties:
      role_claims:
        additionalProperties: true
        type: object
      role_name:
        type: string
    required:
    - role_name
    type: object
...
  handlers.UpdateRoleRequest:
    properties:
      role_claims:
        additionalProperties: true
        type: object
      role_name:
        type: string
    required:
    - role_name
    type: object
...
  repositories.RoleListResult:
    properties:
      limit:
        type: integer
      page:
        type: integer
      roles:
        items:
          $ref: '#/definitions/entities.Role'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the API for Nusarithm IAM Backend
  title: Nusarithm IAM API
  version: "1.0"
paths:
...
  /domains/{domainId}/roles:
    get:
      consumes:
      - application/json
      description: Get all roles for a specific domain
      parameters:
      - description: Domain ID
        in: path
        name: domainId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Role'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get roles by domain
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Domain ID
        in: path
        name: domainId
        required: true
        type: string
      - description: Role data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a role
      tags:
      - roles
...
  /roles:
    get:
      consumes:
      - application/json
      description: Get roles with pagination and search
      parameters:
      - description: Domain ID to filter roles
        in: query
        name: domainId
        type: string
      - description: Search term for role name
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repositories.RoleListResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List roles with pagination
      tags:
      - roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a role
      tags:
      - roles
    get:
      consumes:
      - application/json
      description: Get role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a role
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Update role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Role data
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a role
      tags:
      - roles
...
swagger: "2.0"